Annotate images with current time
And start off at the latest image

+++ camera.py
@@ -4,6 +4,7 @@ import datetime
 import logging
 import os
 import picamera
+import re
 import time
 
 
@@ -13,20 +14,30 @@ base_dir = "/home/pi/camera-images"
 latest = os.path.join(base_dir, "latest.jpg")
 
 def filenames(count):
-    i = 0
+    if os.path.exists(latest):
+        current = os.path.basename(os.path.realpath(latest))
+        m = re.match(r'image(\d+).jpg', current)
+        i = (int(m.group(1)) + 1) % count
+    else:
+        i = 0
+    logging.info("Current image id = {}".format(i))
     while True:
-        yield "{}/image{}.jpg".format(base_dir, i)
+        yield "{}/image{:03}.jpg".format(base_dir, i)
         i += 1
         i %= count
 
 with picamera.PiCamera() as cam:
+    cam.annotate_text_size = 15
     for fn in filenames(50):
         if datetime.time(1,00) <= datetime.datetime.now().time() <= datetime.time(7,00):
             time.sleep(120);
         cam.led = True
-        logging.info("{}: Capturing {}".format(time.strftime("%Y-%m-%d %H:%M:%S"), fn))
+        time_str = time.strftime("%Y-%m-%d %H:%M:%S")
+        cam.annotate_text = time_str
+        logging.info("{}: Capturing {}".format(time_str, fn))
         cam.capture(fn)
         cam.led = False
-        os.remove(latest)
+        if os.path.exists(latest):
+            os.remove(latest)
         os.symlink(fn, latest)
         time.sleep(120)
